+++ plugins/newspack-listings/CHANGELOG.md
@@ -1,3 +1,17 @@
+## newspack-listings [3.7.5](https://github.com/Automattic/newspack-workspace/compare/newspack-listings@3.7.4...newspack-listings@3.7.5) (2026-08-31)
+
+
+### Dependencies
+
+* **newspack-components:** upgraded to 4.7.0
+
+## newspack-listings [3.7.4](https://github.com/Automattic/newspack-workspace/compare/newspack-listings@3.7.3...newspack-listings@3.7.4) (2026-08-20)
+
+
+### Dependencies
+
+* **newspack-components:** upgraded to 4.6.3
+
 ## newspack-listings [3.7.3](https://github.com/Automattic/newspack-workspace/compare/newspack-listings@3.7.2...newspack-listings@3.7.3) (2026-08-19)